requirements.txt
Flask
Flask-CORS
Pillow
tensorflow